+++ app/src/test/groovy/org/thirtyninealpharesearch/chemin/pds3/LabelTest.groovy
@@ -115,7 +115,7 @@ class LabelTest extends Specification {
         def intensity = fields.get 1
         intensity.getName() == "INTENSITY"
         intensity.getDataType() == "ASCII_Real"
-        intensity.getUnit() == "COUNTS"
+        intensity.getUnit() == "COUNT"
         intensity.getBytes() == 7
         intensity.getFormat() == "F7.0"
         intensity.getDescription() == "The intensity of the diffraction for each 2-theta value in column 1"